Refactor: move `Iterator` class to separate files and update `BinaryTree` class
- Created a new header file `tree_iterator.h` for the `Iterator` class definition.
- Created a new implementation file `tree_iterator.tpp` for the `Iterator` class implementation.
- Moved the `Iterator` class definition from `binary_tree.h` to `tree_iterator.h`.
- Moved the `Iterator` class implementation from `binary_tree.tpp` to `tree_iterator.tpp`.
- Updated `binary_tree.h` to include the new `tree_iterator.h` header.
- Updated `binary_tree.tpp` to correctly instantiate the `Iterator` class with the template parameter `T`.
- Ensured that the `Iterator` class works correctly with the `TreeNode<T>` structure.
- Only method `getIterator` remains in `binary_tree.h` to provide an iterator to the root node.
- Minor aesthetic changes to the code for consistency and readability.
